+++ partiql-beamline-cli/src/main.rs
@@ -1,12 +1,14 @@
+mod cli;
+
+use crate::cli::{encode_ion_text, IonPrintMode, DATETIME_FORMAT};
 use clap::{Parser, Subcommand};
 use miette::IntoDiagnostic;
 use partiql_beamline::primitives::{Sample, Tick};
 use partiql_beamline::sim::Sim;
-use partiql_beamline_cliargs::{parse_args, SampleCount, Script, Seed, StartTime};
+use partiql_beamline_cliargs::{parse_args, OutputFormat, SampleCount, Script, Seed, StartTime};
+use partiql_extension_ion::Encoding;
 use std::ops::Add;
 use time::Duration;
-
-const DEFAULT_SAMPLE_COUNT: u64 = 100;
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -30,6 +32,9 @@ pub enum Commands {
 
         #[command(flatten)]
         script: Script,
+
+        #[clap(short = 'f', long = "output-format", value_enum, default_value_t=OutputFormat::Text)]
+        output_format: OutputFormat,
     },
 }
 
@@ -42,33 +47,60 @@ fn main() -> miette::Result<()> {
             seed,
             start_time,
             script,
+            output_format,
         } => {
             let cfg = parse_args(&seed, &start_time).into_diagnostic()?;
-            let t0 = cfg.t0;
-
-            let dt_fmt = time::format_description::well_known::Iso8601::DEFAULT;
-
-            println!("Seed: {}", cfg.seed);
-            println!("Start: {}", cfg.t0.format(&dt_fmt).expect("t0 print"));
-
             let script = script.extract().into_diagnostic()?;
-            let mut sim = Sim::from_config(cfg, script.as_bytes()).into_diagnostic()?;
+            let t0 = cfg.t0;
 
             let sample_count = if let Some(sample_count) = sample_count.sample_count {
                 sample_count
             } else {
-                DEFAULT_SAMPLE_COUNT
+                cli::DEFAULT_SAMPLE_COUNT
             };
 
-            // TODO move to variable iteration as opposed to the current `100` limit
-            for _ in 0..sample_count {
-                if let Ok(Some(Sample {
-                    tick: Tick(t),
-                    value,
-                })) = sim.next_sample()
-                {
-                    let time = t0.add(Duration::milliseconds(t as i64));
-                    println!("[{time}] : {value:?}");
+            match output_format {
+                OutputFormat::Text => {
+                    println!("Seed: {}", cfg.seed);
+                    println!("Start: {}", t0.format(&DATETIME_FORMAT).expect("t0 print"));
+
+                    let mut sim = Sim::from_config(cfg, script.as_bytes()).into_diagnostic()?;
+
+                    for _ in 0..sample_count {
+                        if let Ok(Some(Sample {
+                            tick: Tick(t),
+                            value,
+                        })) = sim.next_sample()
+                        {
+                            let time = t0.add(Duration::milliseconds(t as i64));
+                            println!("[{time}] : {value:?}");
+                        }
+                    }
+                }
+                OutputFormat::Ion => {
+                    let res = encode_ion_text(
+                        IonPrintMode::Compact,
+                        &cli::execute(cfg, script, sample_count)?,
+                        Encoding::Ion,
+                    );
+                    match res {
+                        Ok(out) => println!("{:}", &out),
+                        Err(e) => println!("{:?}", e),
+                    }
+                }
+                OutputFormat::IonPretty => {
+                    let res = encode_ion_text(
+                        IonPrintMode::Pretty,
+                        &cli::execute(cfg, script, sample_count)?,
+                        Encoding::Ion,
+                    );
+                    match res {
+                        Ok(out) => println!("{:}", &out),
+                        Err(e) => println!("{:?}", e),
+                    }
+                }
+                _ => {
+                    todo!("Unsupported output format")
                 }
             }
         }